requirements.txt
anyio==3.6.2
certifi==2022.12.7
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
prettytable==3.5.0
python-telegram-bot==20.0b0
rfc3986==1.5.0
sniffio==1.3.0
wcwidth==0.2.5
